Fix use with bash process substitution and other special files (#88)
* Fix use with bash process substitution and other special files

Using os.path.isfile fails on special files like /dev/fd/N, which bash
uses for process substitution.  It's probably better to assume that
anything that isn't a directory can be treated like a file.

* Add tests

* Remove --show-add-rm option

This was added in #91, but now that /dev/null is treated as a normal
file by default, it's no longer needed.

Simplify test cases accordingly - testing comparisons of /dev/null is
useful, but comparing with and without --show-add-rm is no longer
needed.