Add an environment variable `ICDIFF_DEFAULTS` to allow setting default flags.
I'm not 100% happy with this because there's no way to override boolean
options using `optparse`.  i.e. if `ICDIFF_DEFAULTS=-N` there is no way
to override that on the command line.

Switching to `argparse` would allow that via `BooleanOptionalAction` but
that requires python 3.9.

With `optparse, I could also add `no-XXX` flags for all the existing
boolean flags - LMK.